requirements.txt
python-binance
pandas
requests
